Move NavMenu component to common
