+++ common/src/components/NavMenu/index.ts
@@ -1,0 +1,1 @@
+export { default, NavMenu } from "./NavMenu";
